business_logic_services/vibe_mapping_service/requirements.txt
fastapi
uvicorn
pydantic
httpx
requests
pyyaml
